pyproject.toml
[project]
name = "ai-customer-support"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.6",
    "deepgram-sdk>=6.1.0",
    "dotenv>=0.9.9",
    "elevenlabs>=2.40.0",
    "fastapi>=0.135.2",
    "ffmpeg>=1.4",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "pymysql>=1.1.2",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
]
